requirements.txt
astroid==2.4.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
decorator==5.0.9
Flask==2.0.2
Flask-Cache==0.13.1
Flask-Caching==1.10.1
flask-peewee==3.0.4
Flask-WTF==0.15.1
idna==3.3
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
peewee==3.14.8
pickleshare==0.7.5
prompt-toolkit==3.0.19
Pygments==2.9.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.19.1
pywin32==301
pyzmq==22.1.0
requests==2.26.0
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
wtf-peewee==3.0.2
WTForms==2.3.3
